Paperdoll handler renaming
